IO_2025_26_S/requirements.txt
Flask==3.0.0
matplotlib==3.8.2
python-dotenv==1.0.0
openai==1.54.0